requirements.txt
aioredis==1.2.0
amqp==2.4.2
anyjson==0.3.3
asgi-redis==1.4.3
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.10.1
Automat==0.7.0
autopep8==1.4.4
bcrypt==3.1.4
billiard==3.6.0.0
celery==4.3.0
celery-with-redis==3.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
constantly==15.1.0
cryptography==2.3.1
DBUtils==1.3
defusedxml==0.6.0
Django==2.1
django-cas==2.0.3
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.5.2
django-db==0.0.7
django-mama-cas==2.4.0
django-redis-cache==1.7.1
django-stubs==0.12.1
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.4.2
et-xmlfile==1.0.1
future==0.15.2
hiredis==1.0.0
httplib2==0.9.2
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
Jinja2==2.10
kombu==4.2.1
ldap==1.0.2
ldap3==2.5.1
MarkupSafe==1.1.0
msgpack==0.5.6
msgpack-python==0.5.6
mypy==0.701
mypy-extensions==0.4.1
mysql-replication==0.18
mysqlclient==1.4.2.post1
numpy==1.16.3
odfpy==1.3.6
openpyxl==2.5.3
pandas==0.23.4
paramiko==2.4.1
ply==3.11
pyasn1==0.4.2
pyasn1-modules==0.2.3
pycodestyle==2.5.0
pycparser==2.18
pycryptodomex==3.6.1
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.6.4
PyMySQL==0.9.2
PyNaCl==1.2.1
python-crontab==2.3.5
python-dateutil==2.8.0
python-magic==0.4.15
python-memcached==1.59
pytz==2018.7
PyYAML==3.13
redis==2.10.6
requests==2.20.0
six==1.11.0
sqlparse==0.3.0
tablib==0.12.1
txaio==18.8.1
typed-ast==1.3.5
typing-extensions==3.7.2
unicodecsv==0.14.1
urllib3==1.24.1
vine==1.1.4
webencodings==0.5.1
x509==0.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.6.0
